feat(check-in): share 2-column mood grid between archive card and Check-In tab
Extracts the 2-column mood stacking that lived inline in MoodCardBody
into a shared <MoodGrid> component, then applies it to the live
Check-In tab's Mood quadrant (routes/update/UpdateCheckin.tsx).
Matches the plan's requirement that the layout used in archive cards
("2-column rows rather than all five in one row, odd final emoji
centered in its row") also shows up in the primary check-in view.

MoodGrid is configurable on emoji size / row gap / column gap so both
surfaces pick the dimensions that suit them — archive cards use
size=28 (fits the 120-ish px square body), Mood quadrant uses size=32
(matches the prior single-row quadrant display). The archive card
body now composes MoodGrid in a one-liner, keeping the two surfaces
visually in lockstep if future tweaks land.

Keys use `{emoji}-{position}-{occurrence}` so duplicate emojis at
different indices stay stable-keyed without tripping ESLint's
react/no-array-index-key rule.

npx craco build passes, prettier + eslint clean.
